added lots of stuff

+++ src/mx_atoi.c
@@ -4,7 +4,7 @@ int mx_atoi(char *str) {
     int sign = 1;
     int sum = 0;
 
-    for (;mx_isspace(*str); str++)
+    for (; mx_isspace(*str); str++)
     if (*str == '-' || *str == '+') {
         if (*str == '-')
             sign = -1;
@@ -12,7 +12,6 @@ int mx_atoi(char *str) {
     }
 
     for (; mx_isdigit(*str); str++)
-        sum = 10 * sum + sign * (int) (*str - '0');
+        sum = 10 * sum + sign * (*str - '0');
     return sum;
 }
-
